Separate Search Page into Keyword Seacrh and Category Search
