+++ src/app/search/categorysearch/page.tsx
@@ -1,0 +1,11 @@
+export default function Page() {
+  return (
+    <>
+      <div className="bg-gradient-to-b from-white to-stone-300 h-dvh">
+        <p className="text-gray-700 text-center">
+            Category Page
+        </p>
+      </div>
+    </>
+  );
+}
